requirements.txt
numpy
matplotlib
scikit-learn
Pillow
opencv-python